Improve shortcut behavior on all screen sizes, add shortcut wrapping behavior
This commit makes all parts of the shortcut system much more friendly for non 16:9 aspect ratios. It also raises the limit of allowed shortcuts from 20 to 50, introducing a drawer with wrapping behavior: Shortcuts that do not fit in the shortcut bar can be extended into a drawer.
As such, this commit adds new variables to the CSS sheet, most importantly the max-shortcut-bar-width, which defines which percentage of the screen width may be taken up by the shortcut bar.