Fix touch icon download not working
Previously, icon downloads were essentially downloading cookies, too. In an attempt to fix this, I broke the icon download functionality.
In this commit, we embrace the foreign cookies, as that is but a small price to pay for pretty icons. We will at least, however, not send the referrer in our requests, and as far as I can tell, are not allowing anyone to track the user.