Update script.js
There was a syntax problem in the code preventing several functions from working correctly, after solving it the colors were applied again and the button was opened again, but with one however, when opened it quickly closed by executing the else of the code, when inducing an error in any part of the code works correctly but the rest does not, I still haven't discovered what is causing this. Another problem is that the date is changing every second causing a glitch with languages ​​other than English.